% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDHA.R
\docType{data}
\name{Goolam}
\alias{Goolam}
\title{Goolam}
\format{
An object of class \code{SingleCellExperiment} with 41428 rows and 124 columns.
}
\usage{
Goolam
}
\description{
Goolam dataset in SingleCellExperiment format, include scRNA-seq data and cell type information.
}
\keyword{datasets}
