% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDHA.R
\docType{data}
\name{Goolam_result}
\alias{Goolam_result}
\title{Goolam_result}
\format{
An object of class \code{list} of length 4.
}
\usage{
Goolam_result
}
\description{
Result of processing Goolam dataset using 'scDHA' function.
}
\keyword{datasets}
