% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{scDHA.class}
\alias{scDHA.class}
\title{scDHA classification}
\usage{
scDHA.class(
  train = train,
  train.label = train.label,
  test = test,
  ncores = 15L,
  seed = NULL
)
}
\arguments{
\item{train}{Expression matrix of available data, with rows represent samples and columns represent genes.}

\item{train.label}{A vector containing label for each sample in training data.}

\item{test}{Expression matrix new data for classification, with rows represent samples and columns represent genes.}

\item{ncores}{Number of processor cores to use.}

\item{seed}{Seed for reproducibility.}
}
\value{
A vector contain classified labels for new data.
}
\description{
Perform classification of new data based on available data.
}
