% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDHA.R
\name{scDHA}
\alias{scDHA}
\title{scDHA}
\usage{
scDHA(
  data = data,
  k = NULL,
  sparse = F,
  n = 5000,
  ncores = 15L,
  gen_fil = T,
  do.clus = T,
  sample.prob = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{Gene expression matrix, with rows represent samples and columns represent genes.}

\item{k}{Number of clusters, leave as default for auto detection. Has no effect when \code{do.clus = False}.}

\item{sparse}{Boolen variable indicating whether data is a sparse matrix. The input must be a non negative sparse matrix.}

\item{n}{Number of genes to keep after feature selection step.}

\item{ncores}{Number of processor cores to use.}

\item{gen_fil}{Boolean variable indicating whether to perform scDHA gene filtering before performing dimension deduction and clustering.}

\item{do.clus}{Boolean variable indicating whether to perform scDHA clustering. If \code{do.clus = False}, only dimension deduction is performed.}

\item{sample.prob}{Probability used for classification application only. Leave this parameter as default, no user input is required.}

\item{seed}{Seed for reproducibility.}
}
\value{
List with the following keys:
\itemize{
\item cluster - A numeric vector containing cluster assignment for each sample. If \code{do.clus = False}, this values is always \code{NULL}.
\item latent - A matrix representing compressed data from the input data, with rows represent samples and columns represent latent variables.
}
}
\description{
The main function to perform dimension deduction and clustering.
}
