% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDHA.R
\name{scDHA.w}
\alias{scDHA.w}
\title{scDHA.w}
\usage{
scDHA.w(data = data, sparse = F, ncores = 15L, seed = NULL)
}
\arguments{
\item{data}{Gene expression matrix, with rows represent samples and columns represent genes.}

\item{sparse}{Boolen variable indicating whether data is a sparse matrix. The input must be a non negative sparse matrix.}

\item{ncores}{Number of processor cores to use.}

\item{seed}{Seed for reproducibility.}
}
\value{
A plot with normalized weights of all genes.
}
\description{
This function will plot a graph with normalized weights of all genes so user can select the appropriate number of genes to keep.
}
