% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDHA.R
\docType{data}
\name{Goolam}
\alias{Goolam}
\title{Goolam}
\format{
An object of class \code{list} of length 2.
}
\usage{
Goolam
}
\description{
Goolam dataset in list format, include scRNA-seq data and cell type information.
}
\keyword{datasets}
