% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{scDHA.vis}
\alias{scDHA.vis}
\title{scDHA visulization}
\usage{
scDHA.vis(sc = sc, ncores = 15L, seed = NULL)
}
\arguments{
\item{sc}{Embedding object produced by the \code{scDHA} function.}

\item{ncores}{number of processor cores to use.}

\item{seed}{Seed for reproducibility.}
}
\value{
a list with the following keys:
\itemize{
\item pred - A matrix representing the 2D projection of single-cell data, where rows represent samples and columns represent latent components.
}
}
\description{
Generating 2D embeded data for visulation.
}
